// ApartmentApps.Client.Mobile/src/services/backend/generated/backend.generated.ts
export class QueryResultOfLookupBindingModel { 
    total: number; 
    result: LookupBindingModel[];

    constructor(data?: any) {
        if (data !== undefined) {
            this.total = data["Total"] !== undefined ? data["Total"] : null;
            if (data["Result"] && data["Result"].constructor === Array) {
                this.result = [];
                for (let item of data["Result"])
                    this.result.push(LookupBindingModel.fromJS(item));
            }
        }
    }

    static fromJS(data: any): QueryResultOfLookupBindingModel {
        return new QueryResultOfLookupBindingModel(data);
    }

    toJS(data?: any) {
        data = data === undefined ? {} : data;
        data["Total"] = this.total !== undefined ? this.total : null;
        if (this.result && this.result.constructor === Array) {
            data["Result"] = [];
            for (let item of this.result)
                data["Result"].push(item.toJS());
        }
        return data; 
    }

    toJSON() {
        return JSON.stringify(this.toJS());
    }

    clone() {
        const json = this.toJSON();
        return new QueryResultOfLookupBindingModel(JSON.parse(json));
    }
}

export class LookupBindingModel { 
    id: string; 
    title: string; 
    textPrimary: string; 
    textSecondary: string; 
    imageUrl: string; 
    selected: boolean;

    constructor(data?: any) {
        if (data !== undefined) {
            this.id = data["Id"] !== undefined ? data["Id"] : null;
            this.title = data["Title"] !== undefined ? data["Title"] : null;
            this.textPrimary = data["TextPrimary"] !== undefined ? data["TextPrimary"] : null;
            this.textSecondary = data["TextSecondary"] !== undefined ? data["TextSecondary"] : null;
            this.imageUrl = data["ImageUrl"] !== undefined ? data["ImageUrl"] : null;
            this.selected = data["Selected"] !== undefined ? data["Selected"] : null;
        }
    }

    static fromJS(data: any): LookupBindingModel {
        return new LookupBindingModel(data);
    }

    toJS(data?: any) {
        data = data === undefined ? {} : data;
        data["Id"] = this.id !== undefined ? this.id : null;
        data["Title"] = this.title !== undefined ? this.title : null;
        data["TextPrimary"] = this.textPrimary !== undefined ? this.textPrimary : null;
        data["TextSecondary"] = this.textSecondary !== undefined ? this.textSecondary : null;
        data["ImageUrl"] = this.imageUrl !== undefined ? this.imageUrl : null;
        data["Selected"] = this.selected !== undefined ? this.selected : null;
        return data; 
    }

    toJSON() {
        return JSON.stringify(this.toJS());
    }

    clone() {
        const json = this.toJSON();
        return new LookupBindingModel(JSON.parse(json));
    }
}